Fortschritt meines Abschlussprojektes (Stand: 02.02. 14:30 Uhr)
-> main.kt aufgeräumt,
-> aus variablen konstanten gemacht,
-> variablen/konstanten umbenannt/erneuert,
-> mehr Auswahlmöglichkeiten in Konsole eingebaut
(dementsprechend Code angepasst),
-> Funktionen für fehlerfreien Spielablauf in korrekte Reihenfolge sortiert,
->TODOs geschrieben

morgen:
-> TODOs weiter abarbeiten

+++ src/main/kotlin/Gegner.kt
@@ -1,4 +1,4 @@
-open class Gegner(var name: String, var hp: Double, var schaden: Int) {
+open class Gegner(var name: String, var hp: Double, var schaden: Double) {
 
     var faehigkeiten = mutableListOf<String>(
         "Klon heraufbeschwören",
